fix: 투두 페이지 새로고침 시에 멤버 아이디 설정 안되는 버그 수정
